% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_diagnosis.R
\name{download_diagnosis}
\alias{download_diagnosis}
\title{Download data from Social Diagnosis}
\usage{
download_diagnosis(dir_name = "diagnosis_data")
}
\arguments{
\item{dir_name}{target directory for data.}

\item{language}{\code{PL} for data in Polish or \code{EN} for data in English.}
}
\value{
Invisible 0.
}
\description{
Download data from Social Diagnosis
}
\details{
The size of Social Diagnosis data is around 60 MB.
}

